# perf: Optimize leaf node detection by using direct array length access